Refactor abort_simulation function to return void pointer; update related function signatures and improve sleep timing in monitor thread

+++ src/simulation/philosophers/initialize_philosophers.c
@@ -6,7 +6,7 @@
 /*   By: hsamir <hsamir@student.42kocaeli.com.tr>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/02 17:45:32 by hsamir            #+#    #+#             */
-/*   Updated: 2025/02/05 15:55:50 by hsamir           ###   ########.fr       */
+/*   Updated: 2025/02/06 09:10:54 by hsamir           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ t_philosopher   *initialize_philo(int id, t_simulation *simulation)
             return (free(philo), NULL);
     if (!init_cs(&philo->eat_count, sizeof(int)))
         return (free(philo), NULL);
-    if (!init_cs(&philo->last_meal_time, sizeof(int)))
+    if (!init_cs(&philo->last_meal_time, sizeof(long long)))
         return (abort_cs(&philo->eat_count), free(philo), NULL);
     set_last_meal_time(philo, current_time_ms());
     return (philo);
